Got more progress done.
I'm currently setting up all the basic stuff to get files loaded and
starting to parse the comments. I'm also looking to parse the beginning
line of every handwritten file to match it with it's proper class.

I'm thinking of either using a relative path to a file to link it or
using the class name. Probably for convenience using a class name will
be the simplest for implementation and for users.

After I get all the markdown to generate, I need to figure out how I
should go about transforming it to HTML. I could probably use a library
to do it, but I could leave the option up to the user. This would allow
them to either use a static standard Markdown to HTML generator or use
their own Team custom page layout.